Compilation using skeletons.
