+++ priv/BPF_Boilerplates/Makefile
@@ -5,7 +5,8 @@ SRC_DIR = $(shell realpath .)
 LIBBPF_DIR ?= $(SRC_DIR)/../../benchmarks/libs/libbpf/src
 LIBBPF_INCLUDE ?= $(LIBBPF_DIR)/root/usr/include/
 LIBBPF_HEADERS ?= $(LIBBPF_DIR)/../../headers/
-CFLAGS ?= -I'$(LIBBPF_INCLUDE)' -I'$(LIBBPF_HEADERS)'
+OBJ_FOLDER ?= $(SRC_DIR)/obj/
+CFLAGS ?= -I'$(LIBBPF_INCLUDE)' -I'$(LIBBPF_HEADERS)' -I'$(OBJ_FOLDER)'
 LDFLAGS ?= -L'$(LIBBPF_DIR)'
 LLVM13 ?= /usr/bin
 BPFTOOL ?= $(SRC_DIR)/../../benchmarks/tools/bpftool
