+++ priv/BPF_Boilerplates/OutputFunc.c
@@ -1,0 +1,15 @@
+void output() {
+    int maxPoints = 3;
+    int points = 0;
+    while(1) {
+        fflush(0);
+
+        printf("eBPF program loaded and executing");
+        for (int i = 0; i < (maxPoints + 1); i++) printf((i < points)? ".": " ");
+
+        points = (points + 1) % (maxPoints + 1);
+        printf("\r");
+
+        sleep(1);
+    }
+}
